рабочая версия (удаление списка у пользователя)

+++ pkg/repository/todo_list_postgres.go
@@ -63,10 +63,10 @@ func (r *TodoListPostgres) GetById(userId, listId int) (todo.TodoList, error) {
 	return list, err
 }
 
-func (r *TodoListPostgres) DeleteList(listId int) error {
+func (r *TodoListPostgres) DeleteList(userId, listId int) error {
 	query := fmt.Sprintf(
-		"DELETE FROM %s WHERE id = $1",
-		todoListsTable)
-	_, err := r.db.Exec(query, listId)
+		"DELETE FROM %s tl using %s ul WHERE tl.id = ul.list_id AND ul.user_id=$1 AND ul.list_id=$2",
+		todoListsTable, usersListsTable)
+	_, err := r.db.Exec(query, userId, listId)
 	return err
 }
